move stream module to nested library
