+++ lib/stream.js
@@ -25,11 +25,11 @@ var events = require('events');
 var util = require('util');
 
 util.inherits(Stream, events.EventEmitter);
-Stream.Readable = require('_stream_readable');
-Stream.Writable = require('_stream_writable');
-Stream.Duplex = require('_stream_duplex');
-Stream.Transform = require('_stream_transform');
-Stream.PassThrough = require('_stream_passthrough');
+Stream.Readable = require('stream/readable');
+Stream.Writable = require('stream/writable');
+Stream.Duplex = require('stream/duplex');
+Stream.Transform = require('stream/transform');
+Stream.PassThrough = require('stream/passthrough');
 
 // Backwards-compat with node 0.4.x
 Stream.Stream = Stream;
