+++ lib/stream/duplex.js
@@ -26,8 +26,8 @@
 
 module.exports = Duplex;
 var util = require('util');
-var Readable = require('_stream_readable');
-var Writable = require('_stream_writable');
+var Readable = require('stream/readable');
+var Writable = require('stream/writable');
 
 util.inherits(Duplex, Readable);
