+++ lib/stream/passthrough.js
@@ -25,7 +25,7 @@
 
 module.exports = PassThrough;
 
-var Transform = require('_stream_transform');
+var Transform = require('stream/transform');
 var util = require('util');
 util.inherits(PassThrough, Transform);
