+++ lib/stream/transform.js
@@ -66,7 +66,7 @@
 
 module.exports = Transform;
 
-var Duplex = require('_stream_duplex');
+var Duplex = require('stream/duplex');
 var util = require('util');
 util.inherits(Transform, Duplex);
