+++ node.gyp
@@ -44,11 +44,11 @@
       'lib/readline.js',
       'lib/repl.js',
       'lib/stream.js',
-      'lib/_stream_readable.js',
-      'lib/_stream_writable.js',
-      'lib/_stream_duplex.js',
-      'lib/_stream_transform.js',
-      'lib/_stream_passthrough.js',
+      'lib/stream/readable.js',
+      'lib/stream/writable.js',
+      'lib/stream/duplex.js',
+      'lib/stream/transform.js',
+      'lib/stream/passthrough.js',
       'lib/string_decoder.js',
       'lib/sys.js',
       'lib/timers.js',
@@ -333,7 +333,7 @@
           'action_name': 'node_js2c',
           'inputs': [
             '<@(library_files)',
-            './config.gypi',
+            'config.gypi',
           ],
           'outputs': [
             '<(SHARED_INTERMEDIATE_DIR)/node_natives.h',
